Implement Add button
