+++ Implico/templates/jobDashboardHTML.html
@@ -50,7 +50,7 @@
                       <th class="col-1">Created</th>
                       <th class="col-2">Selected Candidate</th>
                       <th class="col-1"></th>
-                      <th class="col-1"><button type="button" class="btn btn-success">Add</th>
+                      <th class="col-1"><form method="POST"><button type="submit" name="Add" class="btn btn-success">Add</form></th>
                   </tr>
               </thead>
               <tbody>
